Info and dependency views
Rename "version" variables to "alternate".

The data for these two new views are fetched via scripts that are baked
into the binary with some #include and Makefile trickery.

The scripts can be run independently from the `scripts/` folder. They
take a package name as its first argument.

TODO: Tests for the scripts.