Version check functionality
Fetch versions per (filtered) package, and display how many versions of
the currently installed major it is. Also display the latest major, if
any.

If `ENTER` is pressed on a line, the package's version list is show.

Key-bound to `V`.
Initialize with `-V`.

Can be aborted with `ctrl-c`. Blocking of `getch()` is toggled off
during the fetching of all packages, in order to achieve this.

Caveat: if aborted, the alternate window is incomplete, and does not
sync with the package window.

TODO: Cache the result.
TODO: When exiting the version view, display cached packages again.