Add test for the `versions` script
Uses a form of dependency injection, meaning that the actual code needs
to "inject" the string "npm" in `npm.cpp`.

This hopefully opens up for more advanced TUI tests, where `npm` can be
mocked somehow, instead of using the `-f` flag and short-circuiting the
calls to `npm` and returning a fixed set of strings.